숫자 동기화 남은 스킬, 잔탄수 추가

+++ GameClient/D3DBase/XMLReader.h
@@ -12,7 +12,7 @@ public:
 	static void LoadParticle(const wstring& file_name, ParticleSystem* sys);
 private:
 	static void GetNumberUI(CMarkup& xml, UISystem* ui);
-	static const map<wstring, float*>&  GetVariable_map();
-	static map<wstring, float*> variable_map;
+	static const map<wstring, int*>&  GetVariable_map();
+	static map<wstring, int*> variable_map;
 };
